% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ban_nash.R
\name{ban_nash}
\alias{ban_nash}
\title{Optimal bans calculator}
\usage{
ban_nash(W, bans, match_format = c("conquest", "LHS"))
}
\arguments{
\item{W}{Winrate matrix (from the perspective of the Hero)}

\item{bans}{number of bans}

\item{match_format}{Type of match}
}
\value{
A list with following components
\item{bans}{decks banned BY Hero and Opponent}
\item{winrate}{Hero and Opponent's winrate}
\item{stratlist}{Hero's and Opponent's ban options}
\item{conqs}{list of match_format output for each ban combination - first layer is hero's options, second is opponent's}
}
\description{
Finds Nash equilibrium in mixed strategies and expected winrate for both players in a ban phase of a match (Conquest, Conquest where you have to win with all but one deck, LHS)
}
\examples{
ban_nash(W = matrix(runif(16), 4, 4), bans = 1, match_format = "conquest")
}
