% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHS_nash.R
\name{LHS_nash}
\alias{LHS_nash}
\title{LHS Nash calculator}
\usage{
LHS_nash(W)
}
\arguments{
\item{W}{Winrate matrix (from the perspective of the Hero)}
}
\value{
A list of possible state of the game with following parameters (if applicable)
\item{score}{eliminated decks}
\item{havetoplay_hero}{What deck does the hero have to play?}
\item{havetoplay_opp}{What deck does the opponent have to play?}
\item{winrate}{Players' winrate}
\item{nash}{Nash equilibrium in mixed strategies}
\item{game}{payoff matrix for this subgame}
}
\description{
Finds Nash equilibrium in mixed strategies and expected winrate for both players in an LHS match
}
\examples{
LHS_nash(W = matrix(runif(9), 3, 3))
}
