% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialist_run_app.R
\name{run_specialist_app}
\alias{run_specialist_app}
\title{Shiny app for specialist format}
\usage{
run_specialist_app()
}
\value{
shiny.appobj object
}
\description{
Shiny app for finding Nash equilirium strategies
}
