% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_game.R
\name{solve_game}
\alias{solve_game}
\title{zero-sum game solver based on lpSolve}
\usage{
solve_game(W)
}
\arguments{
\item{W}{Winrate matrix (from the perspective of the Hero)}
}
\value{
A list  with a solution
}
\description{
Finds Nash equilibrium in mixed strategies and expected winrate for both players in a zero-sum game specified by a payoff matrix
}
\examples{
solve_game(W = matrix(runif(9), 3, 3))
}
