% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_nash.R
\name{spec_nash}
\alias{spec_nash}
\title{Specialist Nash calculator}
\usage{
spec_nash(W, Hero = 0, Opp = 0, bestof = 5)
}
\arguments{
\item{W}{Winrate matrix (from the perspective of the Hero)}

\item{Hero}{Hero score}

\item{Opp}{Opponent score}

\item{bestof}{Match format.Best-of-n means a player has to win (n+1)/2 games to win the match. Has to be odd.}
}
\value{
A list with components
\item{nash}{nash equilibrium strategies for both players for games after the first one, data frame}
\item{winrates}{winrates for both players, including first game/other games breakdown, matrix}
\item{matchups}{winrate matrix}
\item{scores}{scores named vector}
\item{format}{best of how many games}
}
\description{
Finds Nash equilibrium in mixed strategies and expected winrate for both players in a Specialist match
}
\examples{
spec_nash(W = matrix(runif(9), 3, 3))
}
