% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strike_nash.R
\name{strike_nash}
\alias{strike_nash}
\title{Strike winrate calculator}
\usage{
strike_nash(
  W,
  Hero = 0,
  Opp = 0,
  nash_bans = TRUE,
  nbans = 0,
  HeroBans = NULL,
  OppBans = NULL,
  played = NULL
)
}
\arguments{
\item{W}{Winrate matrix (from the perspective of the Hero)}

\item{Hero}{Hero score}

\item{Opp}{Opponent score}

\item{nash_bans}{whether to find Nash bans, or to to use bans supplied as arguments}

\item{nbans}{number of bans per players. ignored if nash_bans = FALSE}

\item{HeroBans}{coordinate matrix of Hero Bans. ignored if nash_bans = TRUE}

\item{OppBans}{coordinate matrix of Opponent Bans. ignored if nash_bans = TRUE}

\item{played}{coordinate matrix of games that has already been played.}
}
\value{
A list with components
\item{winrates}{A vector of winrates of both players}
\item{matchups}{Winrate matrix}
\item{hero_bans}{Hero bans, whether optimal or supplied by the user}
\item{played}{Games that have already been played, supplied by the user}
\item{isnash}{Whether bans were chosen by the function}
}
\description{
Finds expected winrate for both players in a Strike match
}
\examples{
strike_nash(W = matrix(runif(9), 3, 3))
}
