% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ban_nash.R
\name{ban_nash}
\alias{ban_nash}
\title{Optimal bans calculator}
\usage{
ban_nash(
  W,
  bans,
  nash_fun = c("conquest_nash", "conquest_nash_short", "LHS_nash"),
  shield_h = NULL,
  shield_o = NULL
)
}
\arguments{
\item{W}{Winrate matrix (from the perspective of the Hero)}

\item{bans}{number of bans}

\item{nash_fun}{Type of match}

\item{shield_h}{decks shielded FROM Hero's bans}

\item{shield_o}{decks shielded FROM Opponent's bans}
}
\value{
A list with following components
\item{Hero_bans}{Hero bans}
\item{Opp_bans}{Opponent bans}
\item{V_hero}{Hero winrate}
\item{V_opp}{Hero winrate}
\item{stratlist}{Hero's and Opponent's ban options}
\item{conqs}{list of nash_fun output for each ban combination}
}
\description{
Finds Nash equilibrium in mixed strategies and expected winrate for both players in a ban phase of a match (Conquest, Conquest where you have to win with all but one deck, LHS)
}
\examples{
ban_nash(W = matrix(runif(16), 4, 4), bans = 1, nash_fun = "conquest_nash")
}
