% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_nash.R
\name{shield_nash}
\alias{shield_nash}
\title{Optimal bans calculator}
\usage{
shield_nash(
  W,
  bans,
  nash_fun = c("conquest_nash", "conquest_nash_short", "LHS_nash"),
  shield
)
}
\arguments{
\item{W}{Winrate matrix (from the perspective of the Hero)}

\item{bans}{number of bans}

\item{nash_fun}{Type of match}

\item{shield}{number of shields per player}
}
\value{
A list with following components
\item{Hero_shield}{Hero's optimal shielding strategy}
\item{Opp_shield}{Opponent's optimal shielding strategy}
\item{V_hero}{Hero winrate}
\item{V_opp}{Hero winrate}
\item{stratlist}{Hero's and Opponent's shield options}
\item{conqs}{list of ban_nash output for each shield combination}
}
\description{
Finds Nash equilibrium in mixed strategies and expected winrate for both players in a shield phase of a match (Conquest, Conquest where you have to win with all but one deck, LHS)
}
\examples{
shield_nash(W = matrix(runif(16), 4, 4), bans = 1, nash_fun = "conquest_nash_short", shield = 1)
}
